requirements.txt
# Core Python
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0

# Exchange & Market Data
ccxt>=4.2.0
numpy>=1.24.0

# LLM & API Clients
httpx>=0.26.0

# Database
supabase>=2.3.0
redis>=5.0.1

# API & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dateutil>=2.8.2
aiofiles>=23.2.1

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
hypothesis>=6.92.0

# Data Processing & Storage
pandas>=2.1.0
pyarrow>=14.0.0
matplotlib>=3.8.0

# ML & AI
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Optional — uncomment when needed:
# stable-baselines3>=2.2.0
# gymnasium>=0.29.0
# pymc>=5.10.0
# arviz>=0.17.0
